refactor: extras subpackages mcmc, multifidelity, weights behind lazy imports (2b)
Core now installs with exactly jax, jaxlib, numpy, scipy. The research
surfaces move into optional extras with import guards that name the
pip install jaxbo[extra] fix:

- jaxbo/mcmc: MCMCGP and the numpyro machinery ([mcmc] extra);
  jaxbo.mcmc_models stays as a shim
- jaxbo/multifidelity: the 6 MF + 2 manifold + GradientGP +
  MultipleIndependentOutputsGP models, nn feature maps, and
  serializable.py (extra has no third-party deps today);
  jaxbo.models.* paths stay as shims
- jaxbo/weights: fit_kernel_density, compute_w_gmm, fit_gmm, and the
  shared fit_gmm_constrained extracted from the two byte-identical
  constrained MF methods ([weighted] extra, scikit-learn + KDEpy);
  GPmodel.fit_gmm and both constrained fit_gmm methods delegate, and
  every LW_* gmm_vars branch imports through the guard

Importance weights normalize in float64 before onp.random.choice, which
rejects float32 normalization error. Docstring and type-hint pass on the
moved model files. tox test envs install extras=all; the dev group
carries the extras deps so uv sync keeps the suite runnable. uv.lock
regenerated.

Refs #27
