+++ tox.ini
@@ -14,6 +14,9 @@ skip_missing_interpreters = true
 [testenv]
 # jax comes from the package pin in pyproject.toml (>=0.6,<0.11); the
 # pre-#43/#44 jax<0.10 and pyDOE==0.3.8 constraints are gone, both merged.
+# extras = all: the suite exercises the [mcmc]/[multifidelity]/[weighted]
+# extras and their compat shims, so the tox env installs them.
+extras = all
 deps =
     pytest
 commands = pytest tests
